objectReport/requirements.txt
Django==2.2.3
gunicorn==19.9.0
djangorestframework==3.7.7
mongoengine==0.17.0
django-rest-framework-mongoengine==3.3.1
